requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.2.0
Automat==22.10.0
black==24.2.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
cryptography==41.0.7
cssselect2==0.7.0
cssutils==2.9.0
cycler==0.12.1
daphne==4.0.0
Django==4.2.7
django-colorfield==0.11.0
django-dotenv==1.4.2
django-environ==0.11.2
django-mathfilters==1.0.0
django-phonenumber-field==7.2.0
django-weasyprint==2.2.2
django-wkhtmltopdf==3.4.0
fonttools==4.47.2
html5lib==1.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.2
mypy-extensions==1.0.0
numpy==1.26.3
oscrypto==1.3.0
packaging==23.2
pathspec==0.12.1
pdfkit==1.0.0
phonenumbers==8.13.26
phonenumberslite==8.13.26
pillow==10.2.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.8.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyOpenSSL==23.3.0
pyparsing==3.1.1
pypdf==4.0.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
reportlab==4.0.9
requests==2.31.0
service-identity==23.1.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
weasyprint==60.2
webcolors==1.13
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.14
zope.interface==6.1
zopfli==0.2.3
